chore(team): split archive manager tests
